requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.20
pillow==11.0.0
numpy==1.26.4